Sign in with facebook
Ugh, I feel dirty adding this, but it's the best way, imo to get the
sort of setup that we'll need to properly implement this. I'm not sure
how we'll handle it when we have an android application handling the
frontend, but for now this should be Good Enough(TM)

While I was doing this refactor, I also moved to Django 1.5's custom
model setup, so that we don't have to do any of this silly 1-to-1
relationships between User and Profile bullshit.