Rip out djfb and put in invitations
This shouldn't be a submodule but we're carrying a custom version right
now anyways.

Basically, this came out of talking in IRC with Adityab:

[21:46] <Adityab> rrix: actually I was thinking that this might have
negative survivor bias, in fact maybe the people who are abstaining from
FB are feeling rather lonely (me for example) might be way more desirous
of a hug
[21:46] <rrix> Adityab: that's a good point
[21:47] <rrix> Adityab: how do you suppose gaining some sort of leverage
against creeps in this case?
[21:48] <rrix> Adityab: This turning in to a platform for PUAs is a bit
of a concern for me if I actually decide to make it a thing past my just
fucking around with django
[21:48] <Adityab> rrix: referrals with email addresses
[21:48] <Adityab> rrix: start with an initial trusted group and
organically grow